tests/wpt/web-platform-tests/navigation-api/navigate-event/signal-abort-intercept.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  let abort_signal;
  let onabort_called = false;
  navigation.onnavigate = t.step_func(e => {
    abort_signal = e.signal;
    abort_signal.onabort = () => onabort_called = true;
    e.intercept();
  });

  await navigation.navigate("?1").finished;
  assert_false(abort_signal.aborted);
  assert_false(onabort_called);
}, "event.intercept() does not signal event.signal");
</script>
